Tecemux class in external file
